// helpers/types.ts
export interface SymbolMap<T> {
  [symbol: string]: T;
}

export type eNetwork = eEthereumNetwork;

export enum eEthereumNetwork {
  coverage = "coverage",
  hardhat = "hardhat",
  localhost = "localhost",
  develop = "develop",
  rinkeby = "rinkeby",
  kovan = "kovan",
  main = "main",
  ropsten = "ropsten",
  goerli = "goerli",
}

export enum ConfettiPools {
  proto = "proto",
}

export enum eContractid {
  MintableERC20 = "MintableERC20",
  MintableERC721 = "MintableERC721",
  LendPoolAddressesProvider = "LendPoolAddressesProvider",
  LendPoolAddressesProviderRegistry = "LendPoolAddressesProviderRegistry",
  LendPoolParametersProvider = "LendPoolParametersProvider",
  LendPoolConfigurator = "LendPoolConfigurator",
  ValidationLogic = "ValidationLogic",
  ReserveLogic = "ReserveLogic",
  NftLogic = "NftLogic",
  GenericLogic = "GenericLogic",
  SupplyLogic = "SupplyLogic",
  BorrowLogic = "BorrowLogic",
  LiquidateLogic = "LiquidateLogic",
  ConfiguratorLogic = "ConfiguratorLogic",
  LendPool = "LendPool",
  LendPoolLoan = "LendPoolLoan",
  ReserveOracle = "ReserveOracle",
  ReserveOracleImpl = "ReserveOracleImpl",
  NFTOracle = "NFTOracle",
  NFTOracleImpl = "NFTOracleImpl",
  Proxy = "Proxy",
  MockChainlinkOracle = "MockChainlinkOracle",
  MockNFTOracle = "MockNFTOracle",
  MockReserveOracle = "MockReserveOracle",
  InterestRate = "InterestRate",
  ConfettiUpgradeableProxy = "ConfettiUpgradeableProxy",
  ConfettiProxyAdminTest = "ConfettiProxyAdminTest",
  ConfettiProxyAdminPool = "ConfettiProxyAdminPool", //LendPool Contracts, etc Oracle(Reserve, NFT)
  ConfettiProxyAdminFund = "ConfettiProxyAdminFund", //Treasury Fundings, etc Collector
  WalletBalanceProvider = "WalletBalanceProvider",
  CToken = "CToken",
  DebtToken = "DebtToken",
  CNFT = "CNFT",
  MockCNFT = "MockCNFT",
  ConfettiProtocolDataProvider = "ConfettiProtocolDataProvider",
  IERC20Detailed = "IERC20Detailed",
  IERC721Detailed = "IERC721Detailed",
  FeeProvider = "FeeProvider",
  WETHGateway = "WETHGateway",
  WETHGatewayImpl = "WETHGatewayImpl",
  WETH = "WETH",
  WETHMocked = "WETHMocked",
  WPUNKSGateway = "WPUNKSGateway",
  WPUNKS = "WPUNKS",
  WPUNKSMocked = "WPUNKSMocked",
  SelfdestructTransferMock = "SelfdestructTransferMock",
  LendPoolImpl = "LendPoolImpl",
  LendPoolConfiguratorImpl = "LendPoolConfiguratorImpl",
  LendPoolLoanImpl = "LendPoolLoanImpl",
  CNFTRegistry = "CNFTRegistry",
  CNFTRegistryImpl = "CNFTRegistryImpl",
  CryptoPunksMarket = "CryptoPunksMarket",
  WrappedPunk = "WrappedPunk",
  PunkGateway = "PunkGateway",
  PunkGatewayImpl = "PunkGatewayImpl",
  MockIncentivesController = "MockIncentivesController",
  UIPoolDataProvider = "UIPoolDataProvider",
  ConfettiCollector = "ConfettiCollector",
  ConfettiCollectorImpl = "ConfettiCollectorImpl",
  TimelockControllerFast = "TimelockControllerFast",
  TimelockControllerSlow = "TimelockControllerSlow",
}
// ...
export interface INftParams extends INftBuyItNowParams, INftCollateralParams {
  cNFTImpl: eContractid;
}
// ...
export interface INftCollateralParams {
  baseLTVAsCollateral: string;
  liquidationThreshold: string;
  liquidationBonus: string;
}

export interface INftBuyItNowParams {
  redeemDuration: string;
  buyItNowDuration: string;
  redeemFine: string;
  redeemThreshold: string;
  minBuyItNowFine: string;
}